+++ src/server/socket-server/world/world.ts
@@ -6,7 +6,7 @@ import { WaddleName, WADDLE_ROOMS } from "@server/game-logic/waddles";
 import { PenguinMessenger } from "@server/handlers/messenger";
 import { SettingsManager } from "@server/settings";
 import { GameData } from "@server/timelines/game-data";
-import { ClientSocket } from "..";
+import { ClientSocket } from "../socket-server";
 import { SledRace } from "./sled";
 import { WaddleGame } from "./waddle-game";
 import { WaddleRoom } from "./waddle-room";
